minor changes on algorithm

+++ hackerrank-java/Cakewalk.java
@@ -56,7 +56,8 @@ public class Cakewalk {
 static long marcsCakewalk(Integer[] calorie) {
     Arrays.sort(calorie);
     System.out.println(Arrays.toString(calorie));
-    int minMiles = 0;
+    //note, minMiles has to be data type long, not just because the return type is long, but if you leave minMiles as type int, it won't match some of the larger test cases
+    long minMiles = 0;
     for (int i = calorie.length -1; i >= 0; i--) {
         System.out.println(calorie[i] + " " + (calorie.length-1-i));
         minMiles += calorie[i]*Math.pow(2, (calorie.length-1-i));
